Updated README.md to include a fix for line endings
Updated the readme to include a fix for windows machines...

On windows, you cannot run manage.py (or other related files) as they are using the wrong file endings. It basically doesn't know where to look.

The fix to this is to add .gitattribute locally and rebuild your project, so that it reverts to standard line-endings on your machine (but not elsewhere)